# khadija/tests/contracts/fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ComplianceIQ Core API",
    "description": "\nThe authoritative Core Platform & Data Service for ComplianceIQ.\n\n**Core owns the truth.** Findings, resources, scans and compliance scores\nare produced and owned here. The AI Service reasons *over* this data via\nthis API and never becomes its owner; the dashboard presents it and must\nkeep working when the AI Service is unavailable.\n\n### Authentication\nEvery `/api/v1` endpoint requires `Authorization: Bearer <JWT>` (RS256).\nThe **tenant is taken from the verified token** \u2014 there is no `tenant_id`\nparameter anywhere in this API, and one cannot be supplied.\n\n### Correlation\nSend `X-Correlation-ID` and it is preserved and echoed; omit it and one\nis generated. It appears on every response, including errors.\n\n### Errors\nEvery error \u2014 including validation and unexpected failures \u2014 returns the\nsame envelope: `{\"error\": {\"code\", \"message\", \"correlation_id\", \"details\"}}`.\nBranch on `code`; it is stable. `message` is for humans.\n\n### Versioning\nBreaking changes require a new prefix (`/api/v2`). Additive fields may\nappear in `v1` responses, so clients must ignore unknown fields.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/findings": {
      "get": {
        "tags": [
          "findings"
        ],
        "summary": "List findings",
        "description": "One page of findings for the **authenticated tenant**.\n\nThe tenant is taken from the verified JWT. There is no `tenant_id` parameter, and supplying one has no effect on scoping.\n\n`status=indeterminate` findings are returned like any other and must not be treated as passes \u2014 they mean the check could not be evaluated.",
        "operationId": "list_findings_api_v1_findings_get",
        "parameters": [
          {
            "name": "framework",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Framework id, e.g. iso_27001.",
              "title": "Framework"
            },
            "description": "Framework id, e.g. iso_27001."
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Severity"
                },
                {
                  "type": "null"
                }
              ],
              "description": "critical | high | medium | low",
              "title": "Severity"
            },
            "description": "critical | high | medium | low"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FindingStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "fail | pass | indeterminate. 'indeterminate' is NOT a pass.",
              "title": "Status"
            },
            "description": "fail | pass | indeterminate. 'indeterminate' is NOT a pass."
          },
          {
            "name": "lifecycle_state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LifecycleState"
                },
                {
                  "type": "null"
                }
              ],
              "description": "open | resolved | reopened | suppressed",
              "title": "Lifecycle State"
            },
            "description": "open | resolved | reopened | suppressed"
          },
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Risk domain, e.g. storage.",
              "title": "Domain"
            },
            "description": "Risk domain, e.g. storage."
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CloudProvider"
                },
                {
                  "type": "null"
                }
              ],
              "description": "aws | azure",
              "title": "Provider"
            },
            "description": "aws | azure"
          },
          {
            "name": "resource_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Id"
            }
          },
          {
            "name": "rule_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rule Id"
            }
          },
          {
            "name": "scan_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict to one scan.",
              "title": "Scan Key"
            },
            "description": "Restrict to one scan."
          },
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "name": "detected_after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Detected After"
            }
          },
          {
            "name": "detected_before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Detected Before"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size. Default 50, maximum 100.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size. Default 50, maximum 100."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Zero-based offset.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Zero-based offset."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "description": "Deterministic orderings only; each has a unique tiebreaker.",
              "default": "detected_at_desc"
            },
            "description": "Deterministic orderings only; each has a unique tiebreaker."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_FindingResource_"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or badly signed token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but lacking the 'reader' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid filter, enum value, or pagination bound.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/findings/ai-contract": {
      "get": {
        "tags": [
          "findings"
        ],
        "summary": "List findings in the frozen AI contract shape",
        "description": "Identical filtering and pagination to `GET /findings`, but each item is the exact 11-field payload `contracts/ai_service` specifies.\n\n**INDETERMINATE findings are omitted** \u2014 that contract's status enum has only `pass` and `fail`, and coercing an unevaluated check into either would be inventing a verdict. `total` still reflects the unfiltered match count, so `len(items)` may be smaller than the page size. Use `GET /findings` if you need to see indeterminate results.",
        "operationId": "list_findings_ai_contract_api_v1_findings_ai_contract_get",
        "parameters": [
          {
            "name": "framework",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Framework id, e.g. iso_27001.",
              "title": "Framework"
            },
            "description": "Framework id, e.g. iso_27001."
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Severity"
                },
                {
                  "type": "null"
                }
              ],
              "description": "critical | high | medium | low",
              "title": "Severity"
            },
            "description": "critical | high | medium | low"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FindingStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "fail | pass | indeterminate. 'indeterminate' is NOT a pass.",
              "title": "Status"
            },
            "description": "fail | pass | indeterminate. 'indeterminate' is NOT a pass."
          },
          {
            "name": "lifecycle_state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LifecycleState"
                },
                {
                  "type": "null"
                }
              ],
              "description": "open | resolved | reopened | suppressed",
              "title": "Lifecycle State"
            },
            "description": "open | resolved | reopened | suppressed"
          },
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Risk domain, e.g. storage.",
              "title": "Domain"
            },
            "description": "Risk domain, e.g. storage."
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CloudProvider"
                },
                {
                  "type": "null"
                }
              ],
              "description": "aws | azure",
              "title": "Provider"
            },
            "description": "aws | azure"
          },
          {
            "name": "resource_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Id"
            }
          },
          {
            "name": "rule_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rule Id"
            }
          },
          {
            "name": "scan_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict to one scan.",
              "title": "Scan Key"
            },
            "description": "Restrict to one scan."
          },
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "name": "detected_after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Detected After"
            }
          },
          {
            "name": "detected_before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Detected Before"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size. Default 50, maximum 100.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size. Default 50, maximum 100."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Zero-based offset.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Zero-based offset."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "description": "Deterministic orderings only; each has a unique tiebreaker.",
              "default": "detected_at_desc"
            },
            "description": "Deterministic orderings only; each has a unique tiebreaker."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_AiFindingContract_"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or badly signed token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but lacking the 'reader' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid filter, enum value, or pagination bound.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/findings/{finding_id}": {
      "get": {
        "tags": [
          "findings"
        ],
        "summary": "Get one finding",
        "operationId": "get_finding_api_v1_findings__finding_id__get",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Physical finding id (per-scan).",
              "title": "Finding Id"
            },
            "description": "Physical finding id (per-scan)."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingResource"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or badly signed token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but lacking the 'reader' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid filter, enum value, or pagination bound.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such finding **in this tenant**. Returned identically whether the finding does not exist or belongs to another tenant, so the response cannot be used to probe for other tenants' data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/findings/{finding_id}/ai-contract": {
      "get": {
        "tags": [
          "findings"
        ],
        "summary": "Get one finding in the frozen AI contract shape",
        "description": "The exact 11-field payload `contracts/ai_service` specifies. Provided so the AI Service can consume a byte-stable shape independent of additive changes to the richer API schema.",
        "operationId": "get_finding_ai_contract_api_v1_findings__finding_id__ai_contract_get",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Finding Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiFindingContract"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or badly signed token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but lacking the 'reader' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid filter, enum value, or pagination bound.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such finding in this tenant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The finding cannot be represented in the AI contract \u2014 it is INDETERMINATE, or its framework/domain is outside the contract's closed vocabulary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CloudProvider": {
        "type": "string",
        "enum": [
          "aws",
          "azure"
        ],
        "title": "CloudProvider",
        "description": "Cloud providers the platform can normalize resources from.\n\nAWS is CURRENT (collectors exist per the blueprint's target\narchitecture); AZURE is DESIGNED \u2014 no collector exists yet, but the\nDomain must already be able to represent an Azure-sourced resource\nwithout a rewrite (blueprint \u00a77, ADR-002)."
      },
...
      "FindingStatus": {
        "type": "string",
        "enum": [
          "fail",
          "pass",
          "indeterminate"
        ],
        "title": "FindingStatus",
        "description": "Mirrors the three-valued outcome a rule evaluation produces\n(:class:`domain.rules.conditions.EvaluationResult`) at the Finding\nlevel: a violation (``FAIL``), a pass (``PASS``), or a result that\ncould not be determined from the data collected (``INDETERMINATE``)."
      },
...
      "LifecycleState": {
        "type": "string",
        "enum": [
          "open",
          "resolved",
          "reopened",
          "suppressed"
        ],
        "title": "LifecycleState",
        "description": "Where a logical finding stands right now.\n\n``REOPENED`` is deliberately distinct from ``OPEN``. Both mean \"the\nissue is live\", but a reopened finding is a *regression* \u2014 something\nthat was fixed and broke again \u2014 which is a materially different\nsignal for a security team than a problem that was never fixed. It\nusually indicates a process failure (drift, an unreviewed rollback,\nIaC not matching reality) rather than an unstarted task."
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "critical",
          "high",
          "medium",
          "low"
        ],
        "title": "Severity",
        "description": "Static severity of a rule violation (\"how serious is this in the\nabstract\") \u2014 distinct from ``RiskScore`` (contextual) and\n``ConfidenceScore`` (data reliability). See blueprint \u00a713.\n\nVocabulary fixed to the four values confirmed by the Core\u2194AI Service\nintegration handoff (no ``INFO``) \u2014 Phase 1's original plan flagged\nthis as an open ambiguity (blueprint \u00a713 gives no enumeration) and\nused a 5-level placeholder; this replaces it with the authoritative\nexternal vocabulary now that it is known, rather than carrying two\ndivergent severity scales."
      },
      "SortOrder": {
        "type": "string",
        "enum": [
          "detected_at_desc",
          "detected_at_asc",
          "severity_desc"
        ],
        "title": "SortOrder",
        "description": "Allowed sort orders for findings.\n\nA closed enum rather than a free \"sort by\" string: arbitrary sort\nfields are both an injection surface and a performance trap (sorting\nby an unindexed column on a large tenant). Every value here is backed\nby an index.\n\nEach order ends with a unique tiebreaker (``id``). Without one,\nrows with equal ``detected_at`` can be returned in a different order\non each query, so a row can appear on two pages or none \u2014 the\nclassic unstable-pagination bug (\u00a719)."
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "RS256 JWT issued by complianceiq-core. Required claims: sub, tenant_id, roles, iss=complianceiq-core, aud=complianceiq, exp. Verify against /.well-known/jwks.json."
      }
    }
  }
}